Rename admin_rpc to adminrpc.
